+++ confgenerator.h
@@ -9,7 +9,7 @@
 
 // Constants
 #define MCCONF_SIGNATURE		3154770096
-#define APPCONF_SIGNATURE		982638166
+#define APPCONF_SIGNATURE		296593100
 
 // Functions
 int32_t confgenerator_serialize_mcconf(uint8_t *buffer, const mc_configuration *conf);