+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 	* Improve stability when not using the image system
 	* Many new extensions and access to more configuration parameters.
 * Added PWM + ABI Inverted encoder mode.
+* App ADC coasting brake support.
 
 ### 7.00
 #### 2026-05-15